Initial MayhemHeroes Integration
