+++ mayhem/Dockerfile
@@ -5,4 +5,4 @@ ADD . /src
 WORKDIR /src
 RUN python3 -m pip install .
 
-CMD ["/src/mayhem/fuzz_lang.py"]
+CMD ["/src/mayhem/fuzz_cal.py"]
