+++ mayhem/Mayhemfile
@@ -1,9 +1,9 @@
-project: sqlglot
-target: fuzz-lang
+project: icalendar
+target: fuzz-cal
 testsuite:
     - file://mayhem/testsuite
 
 cmds:
-  - cmd: /src/mayhem/fuzz_lang.py
+  - cmd: /src/mayhem/fuzz_cal.py
     libfuzzer: true
-    timeout: 20
+    timeout: 12
