+++ mayhem/fuzz_cal.py
@@ -1,0 +1,24 @@
+#!/usr/bin/env python3
+
+import atheris
+import sys
+import fuzz_helpers
+
+with atheris.instrument_imports(include=['icalendar']):
+    from icalendar import Calendar
+
+
+def TestOneInput(data):
+    fdp = fuzz_helpers.EnhancedFuzzedDataProvider(data)
+    try:
+        Calendar.from_ical(fdp.ConsumeRemainingString())
+    except ValueError:
+        return -1
+
+def main():
+    atheris.Setup(sys.argv, TestOneInput)
+    atheris.Fuzz()
+
+
+if __name__ == "__main__":
+    main()
